+++ happypages-app/spec/requests/specs/dashboard_spec.rb
@@ -1,6 +1,8 @@
 require "rails_helper"
 
 RSpec.describe "Specs::Dashboard", type: :request do
+  include ActiveSupport::Testing::TimeHelpers
+
   let(:organisation) { create(:organisation) }
   let(:client) { create(:specs_client, :with_password, organisation: organisation) }
 
@@ -24,6 +26,20 @@ RSpec.describe "Specs::Dashboard", type: :request do
       create(:specs_project, :org_scoped, organisation: other_org, name: "Secret Project")
       get specs_dashboard_path
       expect(response.body).not_to include("Secret Project")
+    end
+
+    it "expires session after 24 hours" do
+      post specs_login_path, params: { email: client.email, password: "SecurePass123!" }
+      get specs_dashboard_path
+      expect(response).to have_http_status(:ok)
+
+      # Simulate 25 hours passing by manipulating the session cookie
+      # We need to make a request that sets specs_last_seen in the past
+      # The simplest approach: travel in time
+      travel 25.hours do
+        get specs_dashboard_path
+        expect(response).to redirect_to(specs_login_path)
+      end
     end
   end
 end